Fix event-stream security vulnerability
`gulp-static-gen@1.0.2` depends upon `gulp-livereload@^3.8.1`, which in
turn depends upon `event-stream@^3.1.7`. Unfortunately
`event-stream@3.3.6` is compromised by a malicious maintainer:
https://github.com/dominictarr/event-stream/issues/116. This commit
fixes the vulnerability by forcing an upgrade of the transitive
dependency `gulp-livereload` to version `4.0.1`, a version that dependes
upon `event-stream@3.3.4` (note the absence of a `^` sign, so it is a
safe version.