chore: use parcel bundler for dev and prod build
This project was using an old version of node and gulp to build it.
Some dependecies were troublesome to install and there was a big
gulpfile to config the build.

Nowadays there are simpler options with less configuration, such as the
parcel bundler.

For these reasons the Node version was update to the latest LTS version
14.5.0. A .nvmrc file was added to keep track of the node version used.
parcel-bundler and its handlebar were added and configured.